training/requirements.txt
pandas
numpy
scikit-learn
joblib
torch
sqlalchemy
psycopg2-binary
python-dotenv